requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-heroku==0.3.1
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
django-social-share==2.3.0
django-taggit==3.1.0
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
psycopg2==2.9.4
psycopg2-binary==2.9.5
pycparser==2.21
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
